requirements.txt
openai
moviepy
pydub
pillow==9.5.0